SimpleButton example created material UI added styled components installed for shared package
